requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
mysql-connector-python==8.0.29
numpy==1.22.4
pandas==1.4.2
protobuf==4.21.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
yarl==1.7.2
